requirements.txt
streamlit
plotly
pandas
scikit-learn
numpy
plotly.express 
pip
matplotlib
sklearn.manifold 
sklearn.decomposition
sklearn.ensemble